LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.baseline.invalid-expected.txt

2d.text.baseline.invalid


FAIL OffscreenCanvas test: 2d.text.baseline.invalid assert_equals: ctx.textBaseline === 'top' (got bogus[string], expected top[string]) expected "top" but got "bogus"
